Update dfx.json and setup scripts for ledger canister configuration
- Changed specified_id in dfx.json from "x5qut-viaaa-aaaar-qajda-cai" to "ryjl3-tyaaa-aaaaa-aaaba-cai" for the ledger canister.
- Updated local setup script to use the new ledger canister ID for transferring ICP.
- Enhanced logging in the request lock API to include request lock key information.
- Added new README files for test organization and setup instructions.
- Removed deprecated callCanisterService and signer.service files to clean up the codebase.
