+++ src/test/README.md
@@ -1,0 +1,11 @@
+# Test folder
+
+This folder included two kind of test
+
+### Picjs tests
+
+-   For testing normal flow
+
+### Local tests
+
+-   For testing request lock
